Extracted resumed_runner. Runner clean
